Restructure the importer code
